refactor (path): déplacer tout le contenu du domaine Course dans Path (#86)
