--- src/Course/Level.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace IncentiveFactory\Domain\Course;
-
-enum Level: int
-{
-    case Easy = 1;
-    case Medium = 2;
-    case Hard = 3;
-}
