+++ src/Path/CompleteCourse/CourseCompleted.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace IncentiveFactory\Domain\Course\CompleteCourse;
+namespace IncentiveFactory\Domain\Path\CompleteCourse;
 
-use IncentiveFactory\Domain\Course\CourseLog;
+use IncentiveFactory\Domain\Path\CourseLog;
 use IncentiveFactory\Domain\Shared\EventDispatcher\Event;
 
 final class CourseCompleted implements Event
