+++ src/Path/CourseLogGateway.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace IncentiveFactory\Domain\Course;
+namespace IncentiveFactory\Domain\Path;
 
-use IncentiveFactory\Domain\Shared\Entity\PathInterface;
 use IncentiveFactory\Domain\Shared\Gateway;
 
 /**
@@ -16,5 +15,5 @@ interface CourseLogGateway extends Gateway
 
     public function complete(CourseLog $courseLog): void;
 
-    public function hasAlreadyBegan(PathInterface $path, Course $course): bool;
+    public function hasAlreadyBegan(Path $path, Course $course): bool;
 }
