+++ src/Path/GetCourseBySlug/CourseSlug.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace IncentiveFactory\Domain\Course\GetCourseBySlug;
+namespace IncentiveFactory\Domain\Path\GetCourseBySlug;
 
 use IncentiveFactory\Domain\Shared\Query\Query;
 use Symfony\Component\Validator\Constraints\NotBlank;
