+++ src/Path/GetCourseBySlug/GetCourseBySlug.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace IncentiveFactory\Domain\Course\GetCourseBySlug;
+namespace IncentiveFactory\Domain\Path\GetCourseBySlug;
 
-use IncentiveFactory\Domain\Course\Course;
-use IncentiveFactory\Domain\Course\CourseGateway;
+use IncentiveFactory\Domain\Path\Course;
+use IncentiveFactory\Domain\Path\CourseGateway;
 use IncentiveFactory\Domain\Shared\Query\QueryHandler;
 
 final class GetCourseBySlug implements QueryHandler
