--- src/Shared/Entity/CourseLogInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace IncentiveFactory\Domain\Shared\Entity;
-
-interface CourseLogInterface
-{
-}
